pyproject.toml
[tool.poetry]
name = "DisBot"
version = "0.1.0"
description = ""
authors = ["Rohit Chaurasia <rohit.chaurasia997@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pylint = "^2.6.0"
discord = "^1.0.1"
"discord.py" = "^1.6.0"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
